Fix outdated code in mstl
unary_function and binary_function were removed in C++17, so cause errors when compiling with the latest clang version. They can usually just be deleted, see e.g.
https://stackoverflow.com/questions/63577103/what-is-an-equivalent-replacement-for-stdunary-function-in-c17
https://devblogs.microsoft.com/cppblog/c17-feature-removals-and-deprecations/#:~:text=if%20you%20have%20classes%20deriving,eliminate%20the%20inheritance

gamma is a non-standard extension in the GNU libc, see https://www.gnu.org/software/libc/manual/html_node/Special-Functions.html#index-gamma
And being non-standard, it may give different results on different systems, see https://stackoverflow.com/questions/18116376/what-is-the-definition-for-gammadouble-x-and-why-is-it-different-on-two-gcc-ve 😱